Added handling in main to load files in directories
