+++ src/cogs/general/general_define.py
