+++ src/cogs/general/general_fun.py
