+++ src/cogs/general/general_utility.py
