+++ src/cogs/logging/logging_member_ban.py
@@ -1,12 +1,12 @@
 """
-Logs when a user is kicked
+Logs member bans
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds import embed_kick
-from ._settings import log_channel, server_info
+from utilities.cog_helpers._embeds import embed_ban
+from cogs._settings import log_channel, server_info
 
 
-class LoggingKicks(commands.Cog):
+class LoggingBans(commands.Cog):
     """
     Simple listener to on_member_remove
     then checks the audit log for exact details
@@ -17,7 +17,7 @@ class LoggingKicks(commands.Cog):
     @commands.Cog.listener()
     async def on_member_remove(self, member):
         """
-        First we don't log kicks for unapproved people.
+        First we don't log bans for unapproved people.
         then we grab the guild, and from there read the last entry in the audit log.
         """
         if 'Needs Approval' in [role.name for role in member.roles]:
@@ -26,10 +26,10 @@ class LoggingKicks(commands.Cog):
         current_guild = self.bot.get_guild(server_info['id'])
         audit_log = [entry async for entry in current_guild.audit_logs(limit=1)][0]
 
-        if str(audit_log.action) == 'AuditLogAction.kick':
+        if str(audit_log.action) == 'AuditLogAction.ban':
             if audit_log.target == member:
 
-                embed = embed_kick(member, audit_log)
+                embed = embed_ban(member, audit_log)
 
                 logs_channel = await self.bot.fetch_channel(log_channel['mod_log'])
                 await logs_channel.send(embed=embed)
@@ -40,5 +40,4 @@ def setup(bot):
     """
     Necessary for loading the cog into the bot instance.
     """
-
-    bot.add_cog(LoggingKicks(bot))
+    bot.add_cog(LoggingBans(bot))
