+++ src/cogs/logging/logging_member_roles.py
@@ -3,7 +3,7 @@ Logging for role changes. Logs the user who did the changing, the target user an
 """
 from discord.ext import commands
 from utilities.cog_helpers._embeds import embed_role_add, embed_role_remove
-from ._settings import log_channel, server_info
+from cogs._settings import log_channel, server_info
