+++ src/cogs/logging/logging_name_changes.py
@@ -3,7 +3,7 @@ logs when a username is changed.
 """
 from discord.ext import commands
 from utilities.cog_helpers._embeds import embed_name_change, embed_verified_success
-from ._settings import log_channel
+from cogs._settings import log_channel
 
 
 class LoggingNameChanges(commands.Cog):
