+++ src/cogs/utility/event_challenges.py
